requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pkg-resources==0.0.0
py-sds011==0.9
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pytz==2020.5
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
waveshare-epd==0.0.0
Werkzeug==1.0.1
